DataCrawler/requirements.txt
selenium==4.15.0
pyodbc==4.0.39
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
cryptography==41.0.7